PWN-4941 - Remove BiometricsFragment.kt (#771)
* PWN-4941 - BiometricsFragment.kt removed - logic migrated to NewCreatePinFragment.kt

* PWN-4941 - changes according to comments

* PWN-4941 - changes according to comments (Gleb please stop =))
