--- app/src/main/res/layout/fragment_biometrics.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/bg_lime"
-    tools:context=".auth.ui.pin.biometrics.BiometricsFragment">
-
-    <ImageView
-        android:id="@+id/imageViewLock"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:adjustViewBounds="true"
-        android:src="@drawable/ic_lock_new"
-        android:layout_marginHorizontal="40dp"
-        app:layout_constraintBottom_toTopOf="@+id/textViewTitle"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed" />
-
-    <TextView
-        android:id="@+id/textViewTitle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="20dp"
-        android:gravity="center"
-        android:text="@string/auth_create_wallet_use_touch_id_title"
-        android:textAppearance="@style/UiKit.TextAppearance.SemiBold.LargeTitle"
-        app:layout_constraintBottom_toTopOf="@+id/textViewSubtitle"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageViewLock" />
-
-    <TextView
-        android:id="@+id/textViewSubtitle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="16dp"
-        android:gravity="center"
-        android:text="@string/auth_create_wallet_use_touch_id_subtitle"
-        android:textAppearance="@style/UiKit.TextAppearance.Regular.Title3"
-        app:layout_constraintBottom_toTopOf="@+id/linearLayout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textViewTitle" />
-
-    <LinearLayout
-        android:id="@+id/linearLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/bg_buttons_substrate"
-        android:orientation="vertical"
-        android:padding="24dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <org.p2p.uikit.components.UiKitButton
-            android:id="@+id/buttonTouchId"
-            style="@style/UiKit.Components.Button.Large"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:backgroundTint="@color/button_snow"
-            android:text="@string/auth_create_wallet_use_touch_id"
-            android:textColor="@color/text_night"
-            app:icon="@drawable/ic_fingerprint"
-            app:iconGravity="textEnd"
-            app:iconTint="@color/icons_night" />
-
-        <org.p2p.uikit.components.UiKitButton
-            android:id="@+id/buttonDontUse"
-            style="@style/UiKit.Components.Button.Large.Text"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/ui_kit_button_large_height"
-            android:layout_marginTop="12dp"
-            android:text="@string/auth_create_wallet_dont_use"
-            android:textColor="@color/text_lime" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
